# Conflicts:
#	client/Lykke.Service.PayInternal.Client/Lykke.Service.PayInternal.Client.csproj
#	client/Lykke.Service.PayInternal.Client/Models/PaymentRequest/PaymentRequestTransactionModel.cs
#	src/Lykke.Service.PayInternal.AzureRepositories/Transaction/BlockchainTransactionEntity.cs
#	src/Lykke.Service.PayInternal.Core/Domain/Transaction/IBlockchainTransaction.cs
#	src/Lykke.Service.PayInternal.Core/Domain/Transaction/ICreateTransaction.cs
#	src/Lykke.Service.PayInternal.Core/Domain/Transaction/ICreateTransactionRequest.cs
#	src/Lykke.Service.PayInternal.Core/Lykke.Service.PayInternal.Core.csproj
#	src/Lykke.Service.PayInternal.Services/Domain/CreateTransaction.cs
#	src/Lykke.Service.PayInternal.Services/Lykke.Service.PayInternal.Services.csproj
#	src/Lykke.Service.PayInternal/Lykke.Service.PayInternal.csproj
#	src/Lykke.Service.PayInternal/Models/CreateTransactionRequest.cs
#	src/Lykke.Service.PayInternal/Models/WalletStateResponse.cs
#	src/Lykke.Service.PayInternal/Modules/ServiceModule.cs

+++ src/Lykke.Service.PayInternal.Services/Domain/Order.cs
@@ -9,8 +9,8 @@ namespace Lykke.Service.PayInternal.Services.Domain
         public string MerchantId { get; set; }
         public string PaymentRequestId { get; set; }
         public string AssetPairId { get; set; }
-        public double SettlementAmount { get; set; }
-        public double PaymentAmount { get; set; }
+        public decimal SettlementAmount { get; set; }
+        public decimal PaymentAmount { get; set; }
         public DateTime DueDate { get; set; }
         public DateTime CreatedDate { get; set; }
     }
